Pull-card: extract avatar component
You know, for re-use.
